requirements.txt
# Core Deep Learning Framework
torch>=1.9.0
torchvision>=0.10.0

# Image Processing
Pillow>=8.0.0
opencv-python>=4.5.0
albumentations>=1.1.0

# Scientific Computing
numpy>=1.19.0
scipy>=1.7.0

# Visualization and Plotting
matplotlib>=3.3.0
seaborn>=0.11.0

# Progress Bars and Utilities
tqdm>=4.50.0

# Machine Learning Metrics
scikit-learn>=0.24.0
pandas>=1.2.0

# TensorBoard for Training Monitoring
tensorboard>=2.5.0

# Segmentation Models
segmentation-models-pytorch>=0.3.0

# System Monitoring (for efficiency analysis)
psutil>=5.8.0

# Optional: FLOPs Calculation (for efficiency analysis)
# Uncomment if you want to use measure_efficiency.py
# thop>=0.0.31
# ptflops>=0.6.0

# Optional: Model Summary
# torchsummary>=1.5.1
